{{cookiecutter.project_slug}}/pyproject.toml
#[tool.poetry]
# name = "{{cookiecutter.project_slug}}"
# version = "{{cookiecutter.project_version}}"
# description = "{{cookiecotter.project_description}}"
#authors = ["{{cookiecotter.author}}"]
#
#[tool.poetry.dependencies]
#python = "^3.9"
#Django = "^3.2.6"
#djangorestframework = "^3.12.4"
#django-environ-2 = "^2.1.0"
# {%- if cookiecutter.use_sentry == "y" -%}
#sentry-sdk = "^1.3.1"
# {% endif %}

psycopg2-binary = "^2.9.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
bandit = "^1.7.0"
coverage = "^5.5"
django-nose = "^1.4.7"
factory-boy = "^3.2.0"
isort = "^5.9.3"
flake8 = "^3.9.2"
flake8-absolute-import = "^1.0"
flake8-bugbear = "^21.4.3"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.0.0"
flake8-django = "^1.1.2"
flake8-eradicate = "^1.1.0"
flake8-isort = "^4.0.0"
flake8-fixme = "^1.1.1"
flake8-mock = "^0.3"
flake8-multiline-containers = "^0.0.18"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^1.3.0"
flake8-pie = "^0.14.0"
flake8-print = "^4.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-pytest = "^1.3"
flake8-pytest-style = "^1.5.0"
flake8-quotes = "^3.2.0"
flake8-simplify = "^0.14.1"
flake8-todo = "^0.7"
flake8-use-fstring = "^1.1"
flake8-variables-names = "^0.0.4"
flake8-walrus = "^1.1.0"
mypy = "^0.910"
django-stubs = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 170
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
ensure_newline_before_comments = true
line_length = 170
include_trailing_comma = true
extra_standard_library = "nose"
known_django = ["django", "restframework"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = 3.9
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
strict_equality = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true

[tool.mypy.plugins.django-stubs]
django_settings_module = "server.settings"
